+++ packages/jest-runtime/src/cli/index.ts
@@ -8,7 +8,7 @@
 import {cpus} from 'os';
 import * as path from 'path';
 import chalk = require('chalk');
-import {sync as realpath} from 'realpath-native';
+import {sync as _realpath} from 'realpath-native';
 import yargs = require('yargs');
 import type {Config} from '@jest/types';
 import type {JestEnvironment} from '@jest/environment';
@@ -16,9 +16,15 @@ import {CustomConsole} from '@jest/console';
 import {setGlobal} from 'jest-util';
 import {validateCLIOptions} from 'jest-validate';
 import {deprecationEntries, readConfig} from 'jest-config';
+import shouldPreserveSymlinks from 'should-preserve-links';
 import {VERSION} from '../version';
 import type {Context} from '../types';
 import * as args from './args';
+
+const preserveSymlinks = shouldPreserveSymlinks();
+function realpath(p: string) {
+  return preserveSymlinks ? p : _realpath(p);
+}
 
 export async function run(
   cliArgv?: Config.Argv,